protocol: send error message to bad clients
Give older clients a better failure mode since we can't change them.

The problem case is strong-supervisor@1.1.0 will just exit immediately
with no failure when connecting to a newer version. With this behaviour
the client is at least notified there is an error, which will greatly
improve the user experience for any future incarnations of
strongloop/strong-supervisor#73