replace express-ws with internal router
express-ws is built for handling multiple handlers and servers, and
generally includes a bunch of functionality that isn't used.

Additionally, it pulls in and uses an old version of ws which doesn't
play very well with the efforts to reduce the number of copies of ws
are installed (and compiled) when doing `npm install -g strongloop`.

Bumps the major version because it requires an additional argument when
creating a ws-router.