ws: emit connect and connection from channel
Emit 'connect' every time it establishes a connection with 'connect()'
as a client. The event includes the WebSocket instance mainly for
testing purposes.

Emit 'connection' with a WebSocket instance every time a new WS client
connects to our channel. The channel can only be connected to a single
client at a time, so the WebSocket instance is mainly for testing.