ws: fix sending of ack-only messages
The check for whether other messages were queued was inverted. The
result was that as long as we were the last one to send a message, then
we would never even bother to check if we had any outstanding acks to
send. The effect is that which ever side of the channel was the last
one to send a request (responses don't count) was the only one that
initiate a graceful disconnect.