ws: channel tokens only for session reconnect
Modify the protocol so that the channel token is internally allocated on
first connection, and used to identify incoming websockets for the same
channel.

Clients still have tokens, used to authenticate/connect a new channel
request with a specific server entity, and those tokens are public in
the API, persistent, and shared. These are the tokens in the URI used to
connect to a server.