ws: note that server's error events are odd
They are internally used, so they always have a handler. Its harder to
fix than it looks, and central and pm have to handle the error event
anyhow, to it doesn't cause any problems for us, but it means a server
that doesn't handle an error event will never notice, which is not the
expected behaviour of error events.