Accept unauthorized certs for wss://localhost
When connecting over WSS (HTTPS) on localhost (but only on localhost),
then disable TLS certificates checks. This is a short-cut to support
WSS channel when the server has a certificate issued by a non-public
cert authority. @bajtos and @sam-github believe this does not pose
any security risk, because the local machine is fully under our control.